[AssetMapper][FrameworkBundle] Add AssetMapperBundle
The asset mapper services move to the AssetMapper component, which now provides its own
bundle and the "asset_mapper" configuration. The "framework.asset_mapper" configuration
becomes an alias of it.

The section used to read the state of the "assets" and "http_client" sections while loading;
RemoveMissingDependenciesPass now does it on the merged container instead, dropping the asset
package when no default package is registered and reporting the missing HTTP client. The
"cache.asset_mapper" pool travels with the section and is dropped when no "cache.system" pool
is registered. The console commands move to their own file, imported only when the Console
component is installed.
