+++ src/Symfony/Bundle/FrameworkBundle/composer.json
@@ -43,7 +43,7 @@
         "phpstan/phpdoc-parser": "^1.0|^2.0",
         "seld/jsonlint": "^1.10",
         "symfony/asset": "^7.4|^8.0",
-        "symfony/asset-mapper": "^7.4|^8.0",
+        "symfony/asset-mapper": "^8.2",
         "symfony/browser-kit": "^7.4|^8.0",
         "symfony/clock": "^7.4|^8.0",
         "symfony/console": "^8.1.2",
@@ -88,6 +88,7 @@
         "doctrine/persistence": "<1.3",
         "phpdocumentor/reflection-docblock": "<5.2|>=7",
         "phpdocumentor/type-resolver": "<1.5.1",
+        "symfony/asset-mapper": "<8.2",
         "symfony/console": "<8.1.2",
         "symfony/form": "<7.4",
         "symfony/html-sanitizer": "<8.2",
