+++ src/Symfony/Component/AssetMapper/composer.json
@@ -26,15 +26,21 @@
     "require-dev": {
         "symfony/asset": "^7.4|^8.0",
         "symfony/browser-kit": "^7.4|^8.0",
+        "symfony/cache": "^7.4|^8.0",
+        "symfony/config": "^7.4|^8.0",
         "symfony/console": "^7.4|^8.0",
+        "symfony/dependency-injection": "^8.1",
         "symfony/event-dispatcher-contracts": "^3.0",
         "symfony/finder": "^7.4|^8.0",
-        "symfony/framework-bundle": "^7.4|^8.0",
+        "symfony/framework-bundle": "^8.2",
         "symfony/http-foundation": "^7.4|^8.0",
         "symfony/http-kernel": "^7.4|^8.0",
         "symfony/process": "^7.4|^8.0",
         "symfony/runtime": "^7.4|^8.0",
         "symfony/web-link": "^7.4|^8.0"
+    },
+    "conflict": {
+        "symfony/dependency-injection": "<8.1"
     },
     "autoload": {
         "psr-4": { "Symfony\\Component\\AssetMapper\\": "" },
